copy and paste is not a design pattern
Struggling to get the sgx code to compile with a custom target.
Had to use the fortanix vendor name because of libc looking for it *and*
sgx.
turned off the target-family check, not sure it's actually needed to be
turned off but wanted to avoid using the unix file in libc